Organize all protos in mapping in a single folder. (#1147)
