+++ cartographer/mapping/2d/map_limits.h
@@ -23,8 +23,8 @@
 #include "Eigen/Core"
 #include "Eigen/Geometry"
 #include "cartographer/common/math.h"
-#include "cartographer/mapping/2d/proto/map_limits.pb.h"
 #include "cartographer/mapping/2d/xy_index.h"
+#include "cartographer/mapping/proto/2d/map_limits.pb.h"
 #include "cartographer/mapping/trajectory_node.h"
 #include "cartographer/sensor/point_cloud.h"
 #include "cartographer/sensor/range_data.h"
