+++ cartographer/mapping/3d/hybrid_grid.h
@@ -27,8 +27,8 @@
 #include "cartographer/common/make_unique.h"
 #include "cartographer/common/math.h"
 #include "cartographer/common/port.h"
-#include "cartographer/mapping/3d/proto/hybrid_grid.pb.h"
 #include "cartographer/mapping/probability_values.h"
+#include "cartographer/mapping/proto/3d/hybrid_grid.pb.h"
 #include "cartographer/transform/transform.h"
 #include "glog/logging.h"
