+++ cartographer/mapping/internal/2d/local_trajectory_builder_2d.h
@@ -21,13 +21,13 @@
 #include <memory>
 
 #include "cartographer/common/time.h"
-#include "cartographer/mapping/2d/proto/local_trajectory_builder_options_2d.pb.h"
 #include "cartographer/mapping/2d/submap_2d.h"
 #include "cartographer/mapping/internal/2d/scan_matching/ceres_scan_matcher_2d.h"
 #include "cartographer/mapping/internal/2d/scan_matching/real_time_correlative_scan_matcher_2d.h"
 #include "cartographer/mapping/internal/motion_filter.h"
 #include "cartographer/mapping/internal/range_data_collator.h"
 #include "cartographer/mapping/pose_extrapolator.h"
+#include "cartographer/mapping/proto/2d/local_trajectory_builder_options_2d.pb.h"
 #include "cartographer/metrics/family_factory.h"
 #include "cartographer/sensor/imu_data.h"
 #include "cartographer/sensor/internal/voxel_filter.h"
