+++ cartographer/mapping/internal/3d/local_trajectory_builder_3d.h
@@ -21,13 +21,13 @@
 #include <memory>
 
 #include "cartographer/common/time.h"
-#include "cartographer/mapping/3d/proto/local_trajectory_builder_options_3d.pb.h"
 #include "cartographer/mapping/3d/submap_3d.h"
 #include "cartographer/mapping/internal/3d/scan_matching/ceres_scan_matcher_3d.h"
 #include "cartographer/mapping/internal/3d/scan_matching/real_time_correlative_scan_matcher_3d.h"
 #include "cartographer/mapping/internal/motion_filter.h"
 #include "cartographer/mapping/internal/range_data_collator.h"
 #include "cartographer/mapping/pose_extrapolator.h"
+#include "cartographer/mapping/proto/3d/local_trajectory_builder_options_3d.pb.h"
 #include "cartographer/metrics/family_factory.h"
 #include "cartographer/sensor/imu_data.h"
 #include "cartographer/sensor/internal/voxel_filter.h"
