+++ cartographer/mapping/internal/constraints/constraint_builder_3d.h
@@ -35,8 +35,8 @@
 #include "cartographer/mapping/3d/submap_3d.h"
 #include "cartographer/mapping/internal/3d/scan_matching/ceres_scan_matcher_3d.h"
 #include "cartographer/mapping/internal/3d/scan_matching/fast_correlative_scan_matcher_3d.h"
-#include "cartographer/mapping/pose_graph/proto/constraint_builder_options.pb.h"
 #include "cartographer/mapping/pose_graph_interface.h"
+#include "cartographer/mapping/proto/pose_graph/constraint_builder_options.pb.h"
 #include "cartographer/mapping/trajectory_node.h"
 #include "cartographer/metrics/family_factory.h"
 #include "cartographer/sensor/compressed_point_cloud.h"
