+++ cartographer/mapping/internal/optimization/optimization_problem_options.h
@@ -18,7 +18,7 @@
 #define CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_OPTIMIZATION_PROBLEM_OPTIONS_H_
 
 #include "cartographer/common/lua_parameter_dictionary.h"
-#include "cartographer/mapping/pose_graph/proto/optimization_problem_options.pb.h"
+#include "cartographer/mapping/proto/pose_graph/optimization_problem_options.pb.h"
 
 namespace cartographer {
 namespace mapping {
