+++ cartographer/mapping/proto/2d/grid_2d.proto
@@ -1,4 +1,4 @@
-// Copyright 2016 The Cartographer Authors
+// Copyright 2018 The Cartographer Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,13 +14,27 @@
 
 syntax = "proto3";
 
-import "cartographer/mapping/2d/proto/cell_limits.proto";
-import "cartographer/transform/proto/transform.proto";
-
 package cartographer.mapping.proto;
 
-message MapLimits {
-  double resolution = 1;
-  cartographer.transform.proto.Vector2d max = 2;
-  CellLimits cell_limits = 3;
+import "cartographer/mapping/proto/2d/map_limits.proto";
+import "cartographer/mapping/proto/2d/probability_grid.proto";
+
+message Grid2D {
+  message CellBox {
+    int32 max_x = 1;
+    int32 max_y = 2;
+    int32 min_x = 3;
+    int32 min_y = 4;
+  }
+
+  MapLimits limits = 1;
+  // These values are actually int16s, but protos don't have a native int16
+  // type.
+  repeated int32 cells = 2;
+  CellBox known_cells_box = 3;
+  oneof grid {
+    ProbabilityGrid probability_grid_2d = 4;
+  }
+  float min_correspondence_cost = 6;
+  float max_correspondence_cost = 7;
 }
