+++ cartographer/mapping/proto/2d/local_trajectory_builder_options_2d.proto
@@ -18,9 +18,9 @@ package cartographer.mapping.proto;
 
 import "cartographer/mapping/proto/motion_filter_options.proto";
 import "cartographer/sensor/proto/adaptive_voxel_filter_options.proto";
-import "cartographer/mapping/2d/proto/submaps_options_2d.proto";
-import "cartographer/mapping/2d/scan_matching/proto/ceres_scan_matcher_options_2d.proto";
-import "cartographer/mapping/scan_matching/proto/real_time_correlative_scan_matcher_options.proto";
+import "cartographer/mapping/proto/2d/submaps_options_2d.proto";
+import "cartographer/mapping/proto/scan_matching/ceres_scan_matcher_options_2d.proto";
+import "cartographer/mapping/proto/scan_matching/real_time_correlative_scan_matcher_options.proto";
 
 message LocalTrajectoryBuilderOptions2D {
   // Rangefinder points outside these ranges will be dropped.
