+++ cartographer/mapping/proto/2d/map_limits.proto
@@ -14,17 +14,13 @@
 
 syntax = "proto3";
 
-import "cartographer/mapping/2d/proto/grid_2d_options.proto";
-import "cartographer/mapping/2d/proto/probability_grid_range_data_inserter_options_2d.proto";
-import "cartographer/mapping/proto/range_data_inserter_options.proto";
+import "cartographer/mapping/proto/2d/cell_limits.proto";
+import "cartographer/transform/proto/transform.proto";
 
 package cartographer.mapping.proto;
 
-message SubmapsOptions2D {
-  // Number of range data before adding a new submap. Each submap will get twice
-  // the number of range data inserted: First for initialization without being
-  // matched against, then while being matched.
-  int32 num_range_data = 1;
-  GridOptions2D grid_options_2d = 2;
-  RangeDataInserterOptions range_data_inserter_options = 3;
+message MapLimits {
+  double resolution = 1;
+  cartographer.transform.proto.Vector2d max = 2;
+  CellLimits cell_limits = 3;
 }
