+++ cartographer/mapping/proto/2d/submaps_options_2d.proto
@@ -14,25 +14,16 @@
 
 syntax = "proto3";
 
+import "cartographer/mapping/proto/2d/grid_2d_options.proto";
+import "cartographer/mapping/proto/range_data_inserter_options.proto";
+
 package cartographer.mapping.proto;
 
-import "cartographer/mapping/proto/2d/grid_2d.proto";
-import "cartographer/mapping/proto/3d/hybrid_grid.proto";
-import "cartographer/transform/proto/transform.proto";
-
-// Serialized state of a Submap2D.
-message Submap2D {
-  transform.proto.Rigid3d local_pose = 1;
-  int32 num_range_data = 2;
-  bool finished = 3;
-  Grid2D grid = 4;
-}
-
-// Serialized state of a Submap3D.
-message Submap3D {
-  transform.proto.Rigid3d local_pose = 1;
-  int32 num_range_data = 2;
-  bool finished = 3;
-  HybridGrid high_resolution_hybrid_grid = 4;
-  HybridGrid low_resolution_hybrid_grid = 5;
+message SubmapsOptions2D {
+  // Number of range data before adding a new submap. Each submap will get twice
+  // the number of range data inserted: First for initialization without being
+  // matched against, then while being matched.
+  int32 num_range_data = 1;
+  GridOptions2D grid_options_2d = 2;
+  RangeDataInserterOptions range_data_inserter_options = 3;
 }
