+++ cartographer/mapping/proto/3d/local_trajectory_builder_options_3d.proto
@@ -16,11 +16,11 @@ syntax = "proto3";
 
 package cartographer.mapping.proto;
 
+import "cartographer/mapping/proto/3d/submaps_options_3d.proto";
 import "cartographer/mapping/proto/motion_filter_options.proto";
+import "cartographer/mapping/proto/scan_matching/ceres_scan_matcher_options_3d.proto";
+import "cartographer/mapping/proto/scan_matching/real_time_correlative_scan_matcher_options.proto";
 import "cartographer/sensor/proto/adaptive_voxel_filter_options.proto";
-import "cartographer/mapping/scan_matching/proto/real_time_correlative_scan_matcher_options.proto";
-import "cartographer/mapping/3d/proto/submaps_options_3d.proto";
-import "cartographer/mapping/3d/scan_matching/proto/ceres_scan_matcher_options_3d.proto";
 
 // NEXT ID: 18
 message LocalTrajectoryBuilderOptions3D {
