+++ cartographer/mapping/proto/pose_graph_options.proto
@@ -16,8 +16,8 @@ syntax = "proto3";
 
 package cartographer.mapping.proto;
 
-import "cartographer/mapping/pose_graph/proto/constraint_builder_options.proto";
-import "cartographer/mapping/pose_graph/proto/optimization_problem_options.proto";
+import "cartographer/mapping/proto/pose_graph/constraint_builder_options.proto";
+import "cartographer/mapping/proto/pose_graph/optimization_problem_options.proto";
 
 message PoseGraphOptions {
   // Online loop closure: If positive, will run the loop closure while the map
