+++ cartographer/mapping/proto/range_data_inserter_options.proto
@@ -14,7 +14,7 @@
 
 syntax = "proto3";
 
-import "cartographer/mapping/2d/proto/probability_grid_range_data_inserter_options_2d.proto";
+import "cartographer/mapping/proto/2d/probability_grid_range_data_inserter_options_2d.proto";
 
 package cartographer.mapping.proto;
