+++ cartographer/mapping/proto/scan_matching/ceres_scan_matcher_options_2d.proto
