+++ cartographer/mapping/proto/trajectory_builder_options.proto
@@ -15,8 +15,8 @@
 syntax = "proto3";
 
 import "cartographer/transform/proto/transform.proto";
-import "cartographer/mapping/2d/proto/local_trajectory_builder_options_2d.proto";
-import "cartographer/mapping/3d/proto/local_trajectory_builder_options_3d.proto";
+import "cartographer/mapping/proto/2d/local_trajectory_builder_options_2d.proto";
+import "cartographer/mapping/proto/3d/local_trajectory_builder_options_3d.proto";
 
 package cartographer.mapping.proto;
