+++ cartographer/mapping/range_data_inserter_interface.h
@@ -20,8 +20,8 @@
 #include <utility>
 #include <vector>
 
-#include "cartographer/mapping/2d/proto/submaps_options_2d.pb.h"
 #include "cartographer/mapping/grid_interface.h"
+#include "cartographer/mapping/proto/2d/submaps_options_2d.pb.h"
 #include "cartographer/sensor/range_data.h"
 
 namespace cartographer {
